pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nusafetybench"
version = "0.1.0"
description = "Benchmark scaffold for nuclear safety LLM evaluation"
readme = "README.md"
requires-python = ">=3.9"
authors = [
  { name = "TreeEast1" }
]
dependencies = [
  "PyMuPDF>=1.24.0",
  "PyYAML>=6.0.1",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
